requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
boto3==1.42.68
botocore==1.42.68
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.12
django-cors-headers==4.9.0
django-decouple==2.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.15
exceptiongroup==1.3.1
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==25.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jmespath==1.1.0
kombu==5.6.2
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
tomli==2.4.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
zope.interface==8.2
